Crate transaction_parsing

Re-exports

pub use parse_transaction::entry_to_transactions_with_decoding;

Modules

cards

check_signature

dynamic_derivations

parse_transaction

penumbra

penumbra transaction parsing

penumbra_protobuf

Generic Penumbra protobuf decoder using schema

penumbra_schema_parser

Penumbra schema and registry QR parser

zcash

zcash transaction parsing

Enums

Error

Transaction parsing error.

StubNav

Enum describing Stub content. Is used for proper navigation. Variants: AddSpecs (with associated NetworkSpecsKey), LoadMeta (with associated NetworkSpecsKey for the first by order network using those metadata), and LoadTypes

TransactionAction

Enum containing card sets for four different outcomes: importing derivations (Derivations), signing (Sign), accepting (Stub) and reading, for example, in case of an error (Read)

Functions

decode_payload

Decode content of payload enable_dynamic_derivations is a feature flag

prepare_derivations_preview

produce_output

Type Aliases

Result

Transaction parsing result.