transaction_parsing

Module dynamic_derivations

Functions

dd_transaction_msg_genesis_encryption

Partially parses dynamic derivation transaction to make it compatible with multisigner_msg_genesis_encryption

process_dynamic_derivations

Parses dynamic derivation request and returns preview of the result